src/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Session==0.8.0
pymongo==4.10.1
Werkzeug==3.1.3 
gunicorn==23.0.0
python-logging-loki==0.3.1

